moves document specific values to the common document types file CHM-878
